for std vs. theta draw: for article

+++ macro/N15B11/mc_learning/run.sh
@@ -3,7 +3,12 @@
 IONNAME=N15
 BEGANGLE=5
 ENDANGLE=35
-case_n=1
+
+echo -n "Type case number: "
+read case_n
+
+cp -r archive/phi/${IONNAME}/case_${case_n}/ archive/
+
 if [ -d output ];then
 	rm -rf output/*
 else
@@ -18,9 +23,17 @@ else
         mkdir cases/
 fi
 
-rm -fv input/*
+if [ -d input/ ];then
+	rm -fv input/*
+else
+	mkdir input/
+fi
+
 cp archive/case_${case_n}/${IONNAME}/* input/
 
 root -l -b -q "rootFileWriter.C(\"${IONNAME}\", ${case_n}, ${BEGANGLE}, ${ENDANGLE})"
 
 cp -f output/${IONNAME}_case_${case_n}.root cases/
+
+rm -r archive/phi/${IONNAME}/case_${case_n}/ 
+
